include surrogate for PIL
Python3 is very compatible with PIL, provided you toss PIL and use Pillow, instead.

I think we should add a lot more of these hints, because simple things are becoming
show-stoppers in no time.

Remark: This was my first and only test of "caniusepython3", stated a few minutes ago.
I would suggest to do a contest or something to increase the compatibility list in no time.

Remark#2: This project is so awesome! Had you done this a decade before...

Finally: Hi Brett, I'm not sure if I can be there this year. Pretty much weakness of the body...